+++ src/arxiv2md_beta/ir/transforms/__init__.py
@@ -3,7 +3,7 @@
 from arxiv2md_beta.ir.transforms.anchor import AnchorPass  # noqa: F401
 from arxiv2md_beta.ir.transforms.base import IRPass, PassPipeline  # noqa: F401
 from arxiv2md_beta.ir.transforms.figure_reorder import FigureReorderPass  # noqa: F401
-from arxiv2md_beta.ir.transforms.numbering import NumberingPass  # noqa: F401
+from arxiv2md_beta.ir.transforms.numbering import NumberingPass, SectionNumberingPass  # noqa: F401
 from arxiv2md_beta.ir.transforms.section_filter import SectionFilterPass, split_ir_sections  # noqa: F401
 
 __all__ = [
@@ -13,5 +13,6 @@ __all__ = [
     "AnchorPass",
     "SectionFilterPass",
     "FigureReorderPass",
+    "SectionNumberingPass",
     "split_ir_sections",
 ]